adds Routing - adds starter comps for all categories - adds styles for active links
