+++ app/dashboard.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function Dashboard({ title }) {
+  return (
+    <>
+      <p>I AM {title}</p>
+      <p>TODO: {title} go here</p>
+    </>
+  )
+}
